Botones Edita, Eliminar

+++ taekwondo/src/app/components/alumno/alumno.component.ts
@@ -1,10 +1,12 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ɵbypassSanitizationTrustResourceUrl } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
 import { Alumno } from 'src/app/_models/alumno';
 import { AlumnoService } from 'src/app/_services/alumno.service';
 
 import Swal from 'sweetalert2';
+
+declare var $: any;
 
 @Component({
   selector: 'app-alumno',
@@ -18,6 +20,7 @@ export class AlumnoComponent implements OnInit {
   alumnoForm: FormGroup;
   imagen: File;
   submitted = false;
+  modalTitle: String;
 
   constructor(private alumnoService: AlumnoService, private formBuilder: FormBuilder) { }
 
@@ -42,7 +45,8 @@ export class AlumnoComponent implements OnInit {
 
   // Consultar lista de alumnos
   getAlumnos(){
-    this.alumnos = [];
+    this.alumnos = [new Alumno(1, "Kevin", "Villegas", "11-10-1997", "Foto", "Tae Kwon Do", "Cinta Negra", "Seguro M", "Certificado M", "Carta R", "Pass", "Mail"),
+                    new Alumno(2, "Ricardo", "Salvador", "10-11-1997", "Fotografia", "Kick Boxing", "Cinta Morada", "Seguro", "Certificado", "Carta", "Password", "Email")];
     this.alumnoService.getAlumnos().subscribe(
       res => {
         this.alumnos = res;
@@ -65,12 +69,29 @@ export class AlumnoComponent implements OnInit {
 
   // Eliminar un alumno
   deleteAlumno(id){
-    this.alumnoService.deleteAlumno(id).subscribe(
-      res => {
-        this.getAlumnos();
-      },
-      err => console.error(err)
-    )
+    Swal.fire({
+      title: 'Eliminar Alumno',
+      text: '¿Estás seguro que deseas eliminar al alumno?',
+      showDenyButton: false,
+      showCancelButton: true,
+      confirmButtonText: 'Eliminar',
+      denyButtonText: 'No eliminar',
+    }).then((result) => {
+      if(result.isConfirmed){
+        this.alumnoService.deleteAlumno(id).subscribe(
+          res => {
+            Swal.fire(
+              'Eliminado!',
+              'El alumno ha sido eliminado',
+              'success'
+            )
+            $("#alumnoModal").modal("hide");
+            this.getAlumnos();
+          },
+          err => console.error(err)
+        )
+      }
+    });
   }
 
   // Crear un alumno
@@ -82,27 +103,53 @@ export class AlumnoComponent implements OnInit {
       return;
     }
 
-    this.alumnoService.createAlumno(this.alumnoForm.value).subscribe(
-      res => {
-        Swal.fire({
-          position: 'top-end',
-          icon: 'success',
-          title: 'El alumno ha sido creado',
-          showConfirmButton: false,
-          timer: 1500
-        })
-        this.getAlumnos();
-        this.submitted = false;
-      },
-      err => console.error(err)
-    )
+    if(this.modalTitle == "Registrar"){
+      this.alumnoService.createAlumno(this.alumnoForm.value).subscribe(
+        res => {
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'El alumno ha sido registrado',
+            showConfirmButton: false,
+            timer: 1500
+          })
+          $("#alumnoModal").modal("hide");
+          this.getAlumnos();
+          this.submitted = false;
+        },
+        err => console.error(err)
+      )
+    }else{
+      console.log(this.alumnoForm.value);
+      this.alumnoService.updateAlumno(this.alumnoForm.value).subscribe(
+        res => {
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'El alumno ha sido actualizado',
+            showConfirmButton: false,
+            timer: 1500
+          })
+          $("#alumnoModal").modal("hide");
+          this.getAlumnos();
+          this.submitted = false;
+        },
+        err => {
+          console.error(err);
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Error al conectar con el servidor'
+          })
+        }
+      )
+    }
   }
 
   // Actualizar un alumno
   updateAlumno(alumno: Alumno){
     this.submitted = true;
 
-    this.alumnoForm.controls['id_alumno'].setValue(alumno.id_alumno);
     this.alumnoForm.controls['nombre'].setValue(alumno.nombre);
     this.alumnoForm.controls['apellidos'].setValue(alumno.apellidos);
     this.alumnoForm.controls['fecha_nacimiento'].setValue(alumno.fecha_nacimiento);
@@ -112,8 +159,10 @@ export class AlumnoComponent implements OnInit {
     this.alumnoForm.controls['seguro_medico'].setValue(alumno.seguro_medico);
     this.alumnoForm.controls['certificado_medico'].setValue(alumno.certificado_medico);
     this.alumnoForm.controls['carta_responsiva'].setValue(alumno.carta_responsiva);
-    this.alumnoForm.controls['password'].setValue(alumno.password);
     this.alumnoForm.controls['email'].setValue(alumno.email);
+
+    this.modalTitle = alumno.nombre + " " + alumno.apellidos + " (" + alumno.id_alumno + ")";
+    $("#alumnoModal").modal("show");
   }
 
   imagenSelected(event){
@@ -133,4 +182,10 @@ export class AlumnoComponent implements OnInit {
   }
 
   get f() { return this.alumnoForm.controls; }
+
+  openModalAlumno(){
+    this.alumnoForm.reset();
+    this.modalTitle = "Registrar";
+    $("#alumnoModal").modal("show");
+  }
 }
